// src/services/models/ClinicalTeamModel.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ClinicalTeamModel = {
    firstName?: string | null;
    lastName?: string | null;
    email?: string | null;
    phone?: string | null;
    role?: string | null;
    password?: string | null;
    clinicalTeamType?: string | null;
};